fix(chromedriver): version fixes for update, status, and start (#380)
- Set the max versioning set in config.json. This will need to be updated on
every release of chromedriver. This will "fix" chrome and chromedriver
mismatches until Chrome 75 comes out. When it does there will have to be
an update for this again. Possible future work would allow a user
to set this via flag. Example --maxVersions.chrome "74."
- Create a generic way to get a valid version for Chromedriver. If
presented with 2.x, change this to 2.x.0. If presented with a 74.x.x.x,
chop off the last set of numbers and change this to 74.x.x
  - Fixes the status command during a semver check.
  - Fixes the update and start command when starting a specific version of
chromedriver